Correct foldlevel function
The previous one always put lines of a lower fold level into the last
fold.

Also this commit changes the folding to only fold the _contents_ of
diretories.  The directoy name itself is outside of the fold. This makes
a tree with empty and some non-empty directories on the same level more
readable.