Redefine fold mappings to operate on the fold below
Since we now fold the _content_ of directories, but not the directory
entry itself, calling default fold mappings on a directory entry would
operate on the fold the current direction is in, not the content of it
(which would be expected). Therefore those default mappings are changed
to operate on the fold below instead.

This has the drawback that it is not anymore possible to remap those
operations, as they are hardcoded in this plugin now.