Make default options configurable
A new setting 'g:tree_default_options' was introduced to let the user
specify default options. If not set by the user they will be set to
'--dirsfirst --noreport' as was the case before.

The options '-n -F' are separated and always applied as the are required
for vim-tree to work correctly.