Change default mappings for up/down movements
Instead of 'J' an 'K' we now bind it by default to '<C-J>' and '<C-K>'.
This is done since the latter two are used often for up/down movements and
the former ones (especially 'J') are usually non-movements and we should
not introduce a habit of using 'J' for navigation.